+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.7.8
 
-- Increase channel balance reserve to 100,001 sats
+- Respect spendable and receivable limits for max swap calculation 
 - Show peg-in ETA as time and not remaining duration
 
 ## 1.7.7